Add how to deploy keyring on Ubuntu containers
The changes described in this proposal detail how to use the keyring on Linux systems (Ubuntu 18:04). I am requesting this be added to the docs because I spent a long time trying to figure this out and think others can value from this addition. The addition to the docs provides all the necessary prereqs to use keyring in a dockerized environment